Silence -Wsign-compare whining in bsnprintf()
This warning was present since an incorrect cast was removed in r1.11.
Add the cast to the correct place, i.e., cast to the wider type.

ok florian martijn

+++ usr.sbin/ldapd/util.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: util.c,v 1.11 2019/07/03 03:24:03 deraadt Exp $ */
+/*	$OpenBSD: util.c,v 1.12 2019/10/24 12:39:26 tb Exp $ */
 
 /*
  * Copyright (c) 2009 Martin Hedenfalk <martin@bzero.se>
@@ -43,7 +43,7 @@ bsnprintf(char *str, size_t size, const char *format, ...)
 	va_start(ap, format);
 	ret = vsnprintf(str, size, format, ap);
 	va_end(ap);
-	if (ret < 0 || ret >= size)
+	if (ret < 0 || (size_t)ret >= size)
 		return 0;
 
 	return 1;
